requirements.txt
streamlit
stlib
Pillow
pandas
pyarrow
numpy
altair
xgboost
vegafusion
scikit-learn
tabulate
tflite-runtime